feat: added validation results to command and queries

+++ Application/Activities/Command/DeleteActivity.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using Application.Activities.Core;
 using MediatR;
 using Persistence;
 
@@ -9,20 +6,27 @@ namespace Application.Activities.Command
 {
     public class DeleteActivity
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public required string Id { get; set; }
         }
 
-        public class Handler(AppDbContext context) : IRequestHandler<Command>
+        public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
         {
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
-                var activity = await context.Activities.FindAsync([request.Id], cancellationToken) ?? throw new Exception("Activity not found");
+                var activity = await context.Activities.FindAsync([request.Id], cancellationToken);
+
+                if (activity == null) return Result<Unit>.Failure("Activity not found", 404);
 
                 context.Activities.Remove(activity);
 
-                await context.SaveChangesAsync(cancellationToken);
+                var result = await context.SaveChangesAsync(cancellationToken) > 0;
+
+                if (!result) return Result<Unit>.Failure("Failed to delete activity", 400);
+
+                return Result<Unit>.Success(Unit.Value);
+
             }
         }
     }
